Factor NPM dependency detection out into separate module
